+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## 0.4.1
-- Auto update
-- Better Subagent execution
+## 0.4.2
+- Massively improved Linux support
+- Merging and conflict resolution
 
 ## 0.4.0
 - Project templates